# docs/api/admin-api/openapi-full.yaml
...
openapi: 3.0.3
info:
  version: v0.2.3
  title: Private API print-shop
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Andrey Nazarov
  description: |-
    Private REST API проекта print-shop

    ### Пользовательские ограничения полей и ошибки

    #### Ограничения полей
    - required - поле обязательно для заполнения;
    - omitempty - поле может быть не указано (не будет использоваться методом, в который было передано);
    - unique - уникальное значение поля;
    - min=N - поле должно быть не менее N символов;
    - max=N - поле должно быть не более N символов;
    - gte=N - числовое поле должно быть равно или более N;
    - lte=N - числовое поле должно быть равно или менее N;
    - enum - поле должно содержать одно из ENUM значений;
    - UUID - поле формата UUID;
    - pattern=P - поле должно соответствовать регулярному выражению P;

    #### Ошибки
    - ErrVersionInvalid - если передаваемая версия объекта не совпала с текущей версией объекта.\
      Как правило, это означает, что объект был ранее изменён другим процессом;
    - ErrSwitchStatusRejected - перевод в указанный статус объекта отклонён.\
      WorkFlow объекта запрещает переключение в указанный статус;
servers:
  - url: http://localhost:8080/adm
    description: Local Host
...
paths:
...
  /v1/catalog/papers:
    get:
      summary: Метод получения списка бумаги
      description: |-
        Метод предназначен для получения списка всей бумаги, сохраненных в БД

        Фильтр `App.Request.Query.Filter.SearchText` использует для поиска следующие поля:
        - article;
        - caption;

        Возможные значения `App.Request.Query.ListSorter.SortField`:
        - caption,default,asc;
        - createdAt;
        - updatedAt;

        Возможные значения `App.Request.Query.ListPager.PageSize`:
        - от 1 до 1000 (default: 50)

      operationId: getPaperList
      tags:
        - Catalog.Paper

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/App.Request.Query.Filter.SearchText"
        - $ref: "#/components/parameters/Custom.Request.Query.Filter.MaterialTypeIDs"
        - $ref: "#/components/parameters/Catalog.Papers.AdminAPI.Request.Query.Filter.PaperColorIDs"
        - $ref: "#/components/parameters/Catalog.Papers.AdminAPI.Request.Query.Filter.PaperFactureIDs"
        - $ref: "#/components/parameters/Custom.Request.Query.Filter.WidthMin"
        - $ref: "#/components/parameters/Custom.Request.Query.Filter.WidthMax"
        - $ref: "#/components/parameters/Custom.Request.Query.Filter.HeightMin"
        - $ref: "#/components/parameters/Custom.Request.Query.Filter.HeightMax"
        - $ref: "#/components/parameters/Custom.Request.Query.Filter.DensityMax"
        - $ref: "#/components/parameters/Custom.Request.Query.Filter.DensityMin"
        - $ref: "#/components/parameters/App.Request.Query.Filter.Statuses"
        - $ref: "#/components/parameters/App.Request.Query.ListSorter.SortField"
        - $ref: "#/components/parameters/App.Request.Query.ListSorter.SortDirection"
        - $ref: "#/components/parameters/App.Request.Query.ListPager.PageIndex"
        - $ref: "#/components/parameters/App.Request.Query.ListPager.PageSize"

      responses:
        "200":
          description: Получен список бумаги
          content:
            application/json;charset=UTF-8:
              schema:
                  $ref: "#/components/schemas/Catalog.Papers.AdminAPI.Response.Model.PaperList"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"

    post:
      summary: Метод добавления новой бумаги
      description: |-
        Метод предназначен для создания информации о новой бумаге в БД

        **Ограничения и ошибки полей:**
        - article: required,unique,min=3,max=32,pattern=article
        - caption: required,max=64

      operationId: createPaper
      tags:
        - Catalog.Paper

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"

      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/Catalog.Papers.AdminAPI.Request.Model.CreatePaper"

      responses:
        "201":
          description: Информация о бумаге успешно добавлена
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/App.Response.Model.SuccessCreatedItemInt32"
        "400":
          $ref: "#/components/responses/App.ResponseJson.Error400"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"
  /v1/catalog/papers/{paperId}:
    get:
      summary: Метод получения бумаги
      description: Метод предназначен для получения информации о бумаге из БД
      operationId: getPaper
      tags:
        - Catalog.Paper

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/Catalog.Papers.AdminAPI.Request.Path.PaperID"

      responses:
        "200":
          description: Получена информация о бумаге
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/Catalog.Papers.AdminAPI.Response.Model.Paper"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "404":
          $ref: "#/components/responses/App.ResponseJson.Error404"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"

    put:
      summary: Метод сохранения бумаги
      description: |-
        Метод предназначен для сохранения информации о бумаге в БД

        **Ограничения и ошибки полей:**
        - caption: required,max=64

      operationId: storePaper
      tags:
        - Catalog.Paper

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/Catalog.Papers.AdminAPI.Request.Path.PaperID"

      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/Catalog.Papers.AdminAPI.Request.Model.StorePaper"

      responses:
        "204":
          description: Информация о бумаге успешно сохранена
        "400":
          $ref: "#/components/responses/App.ResponseJson.Error400"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "404":
          $ref: "#/components/responses/App.ResponseJson.Error404"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"

    delete:
      summary: Метод удаления бумаги
      description: Метод предназначен для удаления информации о бумаге из БД
      operationId: removePaper
      tags:
        - Catalog.Paper

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/Catalog.Papers.AdminAPI.Request.Path.PaperID"

      responses:
        "204":
          description: Информация о бумаге успешно удалена
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "404":
          $ref: "#/components/responses/App.ResponseJson.Error404"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"
  /v1/catalog/papers/{paperId}/status:
    patch:
      summary: Метод изменения статуса бумаги
      description: |-
        Метод предназначен для изменения статуса бумаги в БД

        **Ограничения и ошибки полей:**
        - tagVersion: required,gte=1 | ErrVersionInvalid
        - status: required,enum | ErrSwitchStatusRejected

      operationId: changePaperStatus
      tags:
        - Catalog.Paper

      security:
        - bearerAuth: []

      parameters:
        - $ref: "#/components/parameters/App.Request.Header.AcceptLanguage"
        - $ref: "#/components/parameters/App.Request.Header.CorrelationID"
        - $ref: "#/components/parameters/Catalog.Papers.AdminAPI.Request.Path.PaperID"

      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/App.Request.Model.ChangeStatus"

      responses:
        "204":
          description: Статус бумаги успешно изменён
        "400":
          $ref: "#/components/responses/App.ResponseJson.Error400"
        "401":
          $ref: "#/components/responses/App.ResponseJson.Error401"
        "403":
          $ref: "#/components/responses/App.ResponseJson.Error403"
        "404":
          $ref: "#/components/responses/App.ResponseJson.Error404"
        "500":
          $ref: "#/components/responses/App.ResponseJson.Error5XX"
...
components:
  parameters:
    App.Request.Header.AcceptLanguage:
      description: https://developer.mozilla.org/ru/docs/Web/HTTP/Headers/Accept-Language
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        default: "ru-RU"
      example: "ru;q=0.9, fr-CH, fr;q=0.8, en;q=0.7, *;q=0.5"
    App.Request.Header.CorrelationID:
      description: Уникальный идентификатор клиента
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      example: "1f0b7925-cee6-4e44-8ea3-4e585d12212d"
    App.Request.Query.Filter.SearchText:
      description: Поисковая строка
      name: query
      in: query
      required: false
      schema:
        type: string
        maxLength: 128
      example: "перфокарта"
    App.Request.Query.Filter.Statuses:
      description: Список статусов через "," (см. App.Enum.Status)
      name: statuses
      in: query
      required: false
      schema:
        type: string
        default: "ENABLED"
      example: "DRAFT,ENABLED"
    App.Request.Query.ListPager.PageIndex:
      description: Индекс страницы списка с записями
      name: pageIndex
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        maximum: 1024
        default: 0
      example: 2
    App.Request.Query.ListPager.PageSize:
      description: Количество записей на страницу списка
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1024
        default: 50
      example: 70
    App.Request.Query.ListSorter.SortField:
      description: Название колонки, по которой происходит сортировка
      name: sortField
      in: query
      required: false
      schema:
        type: string
      example: "caption"
    App.Request.Query.ListSorter.SortDirection:
      description: Направление сортировки
      name: sortDirection
      in: query
      required: false
      schema:
        description: |-
          Направление сортировки:
          - ASC - по возрастанию;
          - DESC - по убыванию;
        type: string
        enum:
          - ASC
          - DESC
        default: "ASC"
      example: "DESC"
    Custom.Request.Query.Filter.DensityMin:
      description: Минимальная плотность (g/m2)
      name: density-min
      in: query
      required: false
      schema:
        type: integer
        maxLength: 10
      example: 100000 # 100 g/m2
    Custom.Request.Query.Filter.DensityMax:
      description: Максимальная плотность (g/m2)
      name: density-max
      in: query
      required: false
      schema:
        type: integer
        maxLength: 10
      example: 100000 # 100 g/m2
...
    Custom.Request.Query.Filter.HeightMin:
      description: Минимальная высота (µm)
      name: height-min
      in: query
      required: false
      schema:
        type: integer
        maxLength: 10
      example: 200000 # 100 mm
    Custom.Request.Query.Filter.HeightMax:
      description: Максимальная высота (µm)
      name: height-max
      in: query
      required: false
      schema:
        type: integer
        maxLength: 10
      example: 200000 # 100 mm
...
    Custom.Request.Query.Filter.MaterialTypeIDs:
      description: Список идентификаторов типов материала через ","
      name: types
      in: query
      required: false
      schema:
        type: string
      example: "1,2,3"
...
    Custom.Request.Query.Filter.WidthMin:
      description: Минимальная ширина (µm)
      name: width-min
      in: query
      required: false
      schema:
        type: integer
        maxLength: 10
      example: 100000 # 100 mm
    Custom.Request.Query.Filter.WidthMax:
      description: Максимальная ширина (µm)
      name: width-max
      in: query
      required: false
      schema:
        type: integer
        maxLength: 10
      example: 100000 # 100 mm
...
    Catalog.Papers.AdminAPI.Request.Path.PaperID:
      description: Идентификатор бумаги
      name: paperId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      example: 1000
    Catalog.Papers.AdminAPI.Request.Query.Filter.PaperColorIDs:
      description: Список идентификаторов цветов бумаги через ","
      name: colors
      in: query
      required: false
      schema:
        type: string
      example: "1,2,3"
    Catalog.Papers.AdminAPI.Request.Query.Filter.PaperFactureIDs:
      description: Список идентификаторов фактур бумаги через ","
      name: factures
      in: query
      required: false
      schema:
        type: string
      example: "1,2,3"
...
  schemas:
    App.Enum.Status:
      description: |-
        Текущий статус записи:
        - DRAFT - черновик;
        - ENABLED - используется;
        - DISABLED - отключена;
      type: string
      enum:
        - DRAFT
        - ENABLED
        - DISABLED
      example: "ENABLED"
    App.Field.Article:
      description: |
        Уникальный артикул записи (любые буквы без пробелов)
        - pattern: ^S+$
      type: string
      # pattern: "^S+$"
      minLength: 2
      maxLength: 64
      example: "item-FLEX-92744"
    App.Field.Caption:
      description: Название записи
      type: string
      maxLength: 128
      example: "Название"
    App.Field.DateTimeCreatedAt:
      description: Дата и время создания записи (UTC)
      type: string
      format: date-time
      example: "2023-01-01T12:00:00.000+02:00"
    App.Field.DateTimeUpdatedAt:
      description: Дата и время обновления записи (UTC)
      type: string
      format: date-time
      example: "2023-01-01T12:00:00.000+02:00"
...
    App.Field.ListPager.Total:
      description: Количество записей в БД с учётом указанных в запросе фильтров
      type: integer
      minimum: 0
      example: 584
...
    App.Field.TagVersion:
      description: Текущая версия записи
      type: integer
      minimum: 1
      example: 5
...
    App.Field.Measure.GramPerMeter2:
      description: Поверхностная плотность (g/m2)
      type: integer
      maxLength: 10
      example: 100000 # 100g/m2
    App.Field.Measure.Micrometer:
      description: Микрометр (µm, 0.001mm)
      type: integer
      maxLength: 10
      example: 100000 # 100mm
    App.Request.Model.ChangeStatus:
      description: |-
        DTO изменения статуса записи

        Возможные варианты переключений:
        - DRAFT -> ENABLED, DISABLED;
        - ENABLED -> DISABLED;
        - DISABLED -> ENABLED;
      type: object
      required:
        - tagVersion
        - status
      properties:
        tagVersion:
          $ref: "#/components/schemas/App.Field.TagVersion"
        status:
          $ref: "#/components/schemas/App.Enum.Status"
...
    App.Response.Model.ErrorList:
      description: DTO списка пользовательских ошибок 400
      type: array
      items:
        $ref: "#/components/schemas/App.Response.Model.ErrorAttribute"
    App.Response.Model.ErrorAttribute:
      description: DTO пользовательской ошибки 400
      type: object
      required:
        - id
        - value
      properties:
        id:
          description: Название атрибута, где произошла ошибка
          type: string
          example: "attributeName"
        value:
          description: Описание ошибки
          type: string
          example: "Атрибут не может быть пустым | Указано недопустимое значение атрибута"
    App.Response.Model.ErrorDetails: # RFC 7807 (Problem Details for HTTP APIs)
      description: DTO следующих ошибок - 401, 403, 404, 5XX
      type: object
      required:
        - title
        - details
        - request
        - time
      properties:
        title:
          description: Краткое описание проблемы, понятное человеку
          type: string
          example: "Resource not found"
        details:
          description: Описание конкретно возникшей ошибки, понятное человеку
          type: string
          example: "Resource [User] with id = [123456] not found. You MUST use PUT to add a resource instead of GET"
        request:
          description: Метод и URL запроса
          type: string
          example: "PUT /users/123456"
        time:
          description: Время возникновения ошибки с точностью до миллисекунд (UTC)
          type: string
          format: date-time
          example: "2023-01-01T12:00:00.000+02:00"
        errorTraceId:
          description: Идентификатор конкретного возникновения ошибки
          type: string
          example: "5add1be1-90ab5d42-02fa8b1f-672503f2"
...
    App.Response.Model.SuccessCreatedItemInt32:
      description: DTO результата создания записи (int32 ID)
      type: object
      required:
        - id
      properties:
        id:
          description: Идентификатор созданной записи
          type: integer
          format: int32
          example: 2147483647
...
    Custom.Enum.PaperSides:
      description: |-
        Стороны бумаги:
        - SAME - одинаковые;
        - DIFFERENT - разные;
      type: string
      enum:
        - SAME
        - DIFFERENT
      example: "SAME"
...
    Custom.Field.Catalog.PaperID:
      description: Идентификатор бумаги (модуль - Catalog.Paper)
      type: integer
      format: int32
      minimum: 1
      example: 2
...
    Custom.Field.Dictionaries.MaterialTypeID:
      description: Идентификатор типа материала (модуль - Dictionaries.MaterialType)
      type: integer
      format: int32
      minimum: 1
      example: 2
    Custom.Field.Dictionaries.PaperColorID:
      description: Идентификатор цвета материала (модуль - Dictionaries.PaperColor)
      type: integer
      format: int32
      minimum: 1
      example: 2
    Custom.Field.Dictionaries.PaperFactureID:
      description: Идентификатор фактуры материала (модуль - Dictionaries.PaperFacture)
      type: integer
      format: int32
      minimum: 1
      example: 2
...
    Catalog.Papers.AdminAPI.Request.Model.CreatePaper:
      description: DTO добавления новой бумаги
      type: object
      required:
        - article
        - caption
        - typeId
        - colorId
        - factureId
        - width
        - height
        - thickness
        - density
        - sides
      properties:
        article:
          $ref: "#/components/schemas/App.Field.Article"
        caption:
          allOf:
            - $ref: "#/components/schemas/App.Field.Caption"
            - maxLength: 64
            - example: "Мелованная"
        typeId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.MaterialTypeID"
        colorId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.PaperColorID"
        factureId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.PaperFactureID"
        width:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        height:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        thickness:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        density:
          $ref: "#/components/schemas/App.Field.Measure.GramPerMeter2"
        sides:
          $ref: "#/components/schemas/Custom.Enum.PaperSides"
    Catalog.Papers.AdminAPI.Request.Model.StorePaper:
      description: DTO сохранения бумаги
      type: object
      required:
        - tagVersion
      properties:
        tagVersion:
          $ref: "#/components/schemas/App.Field.TagVersion"
        article:
          $ref: "#/components/schemas/App.Field.Article"
        caption:
          allOf:
            - $ref: "#/components/schemas/App.Field.Caption"
            - maxLength: 64
            - example: "Мелованная"
        typeId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.MaterialTypeID"
        colorId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.PaperColorID"
        factureId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.PaperFactureID"
        width:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        height:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        thickness:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        density:
          $ref: "#/components/schemas/App.Field.Measure.GramPerMeter2"
        sides:
          $ref: "#/components/schemas/Custom.Enum.PaperSides"
    Catalog.Papers.AdminAPI.Response.Model.PaperList:
      description: DTO - список бумаги
      type: object
      required:
        - items
        - total
      properties:
        items:
          description: Список записей ограниченный фильтром и pageSize
          type: array
          items:
            $ref: "#/components/schemas/Catalog.Papers.AdminAPI.Response.Model.Paper"
        total:
          $ref: "#/components/schemas/App.Field.ListPager.Total"
    Catalog.Papers.AdminAPI.Response.Model.Paper:
      description: DTO - бумага
      type: object
      required:
        - id
        - tagVersion
        - article
        - caption
        - typeId
        - colorId
        - factureId
        - width
        - height
        - thickness
        - density
        - sides
        - status
        - createdAt
      properties:
        id: # paperId
          $ref: "#/components/schemas/Custom.Field.Catalog.PaperID"
        tagVersion:
          $ref: "#/components/schemas/App.Field.TagVersion"
        article:
          $ref: "#/components/schemas/App.Field.Article"
        caption:
          allOf:
            - $ref: "#/components/schemas/App.Field.Caption"
            - maxLength: 64
            - example: "Мелованная"
        typeId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.MaterialTypeID"
        colorId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.PaperColorID"
        factureId:
          $ref: "#/components/schemas/Custom.Field.Dictionaries.PaperFactureID"
        width:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        height:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        thickness:
          $ref: "#/components/schemas/App.Field.Measure.Micrometer"
        density:
          $ref: "#/components/schemas/App.Field.Measure.GramPerMeter2"
        sides:
          $ref: "#/components/schemas/Custom.Enum.PaperSides"
        status:
          $ref: "#/components/schemas/App.Enum.Status"
        createdAt:
          $ref: "#/components/schemas/App.Field.DateTimeCreatedAt"
        updatedAt:
          $ref: "#/components/schemas/App.Field.DateTimeUpdatedAt"
...
  responses:
    App.ResponseJson.Error400:
      description: Некорректные входные данные. Возвращается список атрибутов с ошибками
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorList"
    App.ResponseJson.Error404:
      description: Запрашиваемый ресурс не был найден
      content:
        application/problem+json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorDetails"
    App.ResponseJson.Error5XX:
      description: Внутренняя ошибка сервера
      content:
        application/problem+json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorDetails"
    App.ResponseJson.Error401:
      description: Пользователь не был аутентифицирован
      content:
        application/problem+json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorDetails"
    App.ResponseJson.Error403:
      description: Пользователю с текущими правами доступ отклонён
      content:
        application/problem+json;charset=UTF-8:
          schema:
            $ref: "#/components/schemas/App.Response.Model.ErrorDetails"
  securitySchemes:
    bearerAuth:
      description: https://www.rfc-editor.org/rfc/rfc6750
      type: http
      scheme: bearer
      bearerFormat: JWT